Extract Docker and systemd command builders for task execution
